Create minimal-version
This is a fix that has removed two small sections of code that were messing with the alignment of news feed views. The remaining report link can now be found at the top of an individual's profile page rather than next to every instance of their screen name and picture.